Auth and Account functionalities

+++ src/controllers/accountController.js
@@ -1,4 +1,5 @@
 const Account = require("../../models/index").account;
+const { Op } = require("sequelize");
 exports.createPost = async (req, res, next) => {
   try {
     if (!req.body.name) {
@@ -6,7 +7,18 @@ exports.createPost = async (req, res, next) => {
       error.status = 400;
       throw error;
     }
-    const newAccounts = await Account.create(req.body);
+    const account = { ...req.body, user_id: req.user.id };
+    const nameAlreadyinUse = await Account.findOne({
+      where: {
+        [Op.and]: [{ name: account.name }, { user_id: account.user_id }]
+      }
+    });
+    if (nameAlreadyinUse) {
+      const error = new Error("Conta já cadastrada para este usuário");
+      error.status = 400;
+      throw error;
+    }
+    const newAccounts = await Account.create(account);
     res.status(201).json(newAccounts);
   } catch (error) {
     next(error);
@@ -14,7 +26,10 @@ exports.createPost = async (req, res, next) => {
 };
 exports.getAll = async (req, res, next) => {
   try {
-    const accounts = await Account.findAll({ raw: true });
+    const accounts = await Account.findAll({
+      where: { user_id: req.user.id },
+      raw: true
+    });
     res.status(200).json(accounts);
   } catch (error) {
     next(error);
@@ -38,22 +53,46 @@ exports.updatedAccountByid = async (req, res, next) => {
       error.status = 400;
       throw error;
     }
-    const response = await Account.update(
-      { name: req.body.name },
-      { where: { id: req.params.id } }
-    );
-    res.status(200).json(response[0]);
+    const userAccount = await Account.findOne({
+      where: {
+        [Op.and]: [{ id: req.params.id }, { user_id: req.user.id }]
+      },
+      raw: true
+    });
+
+    if (!userAccount) {
+      const error = new Error("Conta não registrada para usuário atual");
+      error.status = 403;
+      throw error;
+    } else {
+      const response = await Account.update(
+        { name: req.body.name },
+        { where: { id: req.params.id } }
+      );
+      res.status(200).json(response[0]);
+    }
   } catch (error) {
     next(error);
   }
 };
 exports.deleteAccountByid = async (req, res, next) => {
   try {
-    let response;
-    response = await Account.destroy({
-      where: { id: req.params.id }
+    const userAccount = await Account.findOne({
+      where: {
+        [Op.and]: [{ id: req.params.id }, { user_id: req.user.id }]
+      },
+      raw: true
     });
-    res.status(204).json(response);
+    if (!userAccount) {
+      const error = new Error("Conta não registrada para usuário atual");
+      error.status = 403;
+      throw error;
+    } else {
+      let response = await Account.destroy({
+        where: { id: req.params.id }
+      });
+      res.status(204).json(response);
+    }
   } catch (error) {
     next(error);
   }
